+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Molwitch API Changelog
 ## 0.6.1
 1. added new CtTable Clean Rule M  SAL lines with atom positions &lt; 1 are removed from the atom list
+1. MolSearcherFactory now returns an `Optional<MolSearcher>` to support not finding a searcher. Previously would throw NoSuchElementException.
 
 ## 0.6.0
 1. removed `Atom#getRadicalValue()` since we also have `Atom#getRadical()`